tests/wpt/web-platform-tests/geolocation-API/permission.https.html

<!doctype html>
<meta charset=utf-8>
<title>Test geolocation is a powerful feature via Permissions API</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>
promise_test(async (test) => {
  const status = await navigator.permissions.query({ name: "geolocation" });
  assert_true(status instanceof PermissionStatus);
  assert_equals(status.name, "geolocation", `permission's name must be "geolocation"`);
  assert_equals(status.state, "prompt", `permission's state must be "prompt" by default`);
}, `Query "geolocation" powerful feature`);
</script>
